+++ tools/src/tester/StoryEvaluator.ts
@@ -8,7 +8,7 @@
 */
 
 import { ChapterRequest, Parameter, SupplementalChapter } from './types/story.types'
-import { type StoryFile, type ChapterEvaluation, Result, type StoryEvaluation, OutputReference } from './types/eval.types'
+import { type StoryFile, type ChapterEvaluation, Result, type StoryEvaluation } from './types/eval.types'
 import type ChapterEvaluator from './ChapterEvaluator'
 import { overall_result } from './helpers'
 import { StoryOutputs } from './StoryOutputs'
@@ -17,6 +17,7 @@ import { ChapterOutput } from './ChapterOutput'
 import * as semver from '../_utils/semver'
 import _ from 'lodash'
 import { ParsedChapter, ParsedStory } from './types/parsed_story.types'
+import { OutputReference } from './OutputReference'
 
 export default class StoryEvaluator {
   private readonly _chapter_evaluator: ChapterEvaluator
@@ -217,10 +218,7 @@ export default class StoryEvaluator {
   static #extract_params_variables(parameters: Record<string, Parameter>, variables: Set<OutputReference>): void {
     Object.values(parameters ?? {}).forEach((param) => {
       if (typeof param === 'string') {
-        const ref = OutputReference.parse(param)
-        if (ref) {
-          variables.add(ref)
-        }
+        OutputReference.parse(param).forEach((ref) => variables.add(ref))
       }
     })
   }
@@ -229,10 +227,7 @@ export default class StoryEvaluator {
     const request_type = typeof request
     switch (request_type) {
       case 'string': {
-        const ref = OutputReference.parse(request as string)
-        if (ref !== undefined) {
-          variables.add(ref)
-        }
+        OutputReference.parse(request as string).forEach((ref) => variables.add(ref))
         break
       }
       case 'object': {