+++ tools/src/tester/types/eval.types.ts
@@ -83,20 +83,3 @@ export enum Result {
   SKIPPED = 'SKIPPED',
   ERROR = 'ERROR',
 }
-
-export class OutputReference {
-  chapter_id: string
-  output_name: string
-  private constructor (chapter_id: string, output_name: string) {
-    this.chapter_id = chapter_id
-    this.output_name = output_name
-  }
-
-  static parse (str: string): OutputReference | undefined {
-    if (str.startsWith('${') && str.endsWith('}')) {
-      const spl = str.slice(2, -1).split('.', 2)
-      return { chapter_id: spl[0], output_name: spl[1] }
-    }
-    return undefined
-  }
-}